move GeometryCollection to separate file
