+++ doc/symbolic.rst
@@ -13,9 +13,6 @@ Binding an operator to a discretization
 
 .. currentmodule:: pytential
 
-.. autoclass:: GeometryCollection
-.. autofunction:: add_geometry_to_collection
-
 .. autofunction:: bind
 
 PDE operators
