+++ pytential/symbolic/dof_connection.py
@@ -212,7 +212,7 @@ class DOFConnection:
 
 def connection_from_dds(places, from_dd, to_dd):
     """
-    :arg places: a :class:`~pytential.symbolic.execution.GeometryCollection`
+    :arg places: a :class:`~pytential.collection.GeometryCollection`
         or an argument taken by its constructor.
     :arg from_dd: a descriptor for the incoming degrees of freedom. This
         can be a :class:`~pytential.symbolic.primitives.DOFDescriptor`
