+++ pytential/symbolic/execution.py
@@ -23,7 +23,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from typing import Any, Dict, Hashable, Optional
+from typing import Optional
 
 from pymbolic.mapper.evaluator import (
         EvaluationMapper as PymbolicEvaluationMapper)
@@ -546,7 +546,7 @@ class MatVecOp:
 def _prepare_domains(nresults, places, domains, default_domain):
     """
     :arg nresults: number of results.
-    :arg places: a :class:`~pytential.GeometryCollection`.
+    :arg places: a :class:`~pytential.collection.GeometryCollection`.
     :arg domains: recommended domains.
     :arg default_domain: default value for domains which are not provided.
 
@@ -573,9 +573,9 @@ def _prepare_auto_where(auto_where, places=None):
     """
     :arg auto_where: a 2-tuple, single identifier or `None` used as a hint
         to determine the default geometries.
-    :arg places: a :class:`GeometryCollection`,
-        whose :attr:`GeometryCollection.auto_where` is used by default if
-        provided and `auto_where` is `None`.
+    :arg places: a :class:`pytential.collection.GeometryCollection`,
+        whose :attr:`pytential.collection.GeometryCollection.auto_where` is
+        used by default if provided and `auto_where` is `None`.
     :return: a tuple ``(source, target)`` of
         :class:`~pytential.symbolic.primitives.DOFDescriptor`s denoting
         the default source and target geometries.
@@ -598,7 +598,7 @@ def _prepare_auto_where(auto_where, places=None):
 
 def _prepare_expr(places, expr, auto_where=None):
     """
-    :arg places: :class:`~pytential.GeometryCollection`.
+    :arg places: :class:`~pytential.collection.GeometryCollection`.
     :arg expr: a symbolic expression.
     :return: processed symbolic expressions, tagged with the appropriate
         `where` identifier from places, etc.
@@ -621,372 +621,6 @@ def _prepare_expr(places, expr, auto_where=None):
     expr = InterpolationPreprocessor(places)(expr)
 
     return expr
-
-# }}}
-
-
-# {{{ geometry collection
-
-def _is_valid_identifier(name):
-    import keyword
-    return name.isidentifier() and not keyword.iskeyword(name)
-
-
-class _GeometryCollectionDiscretizationCacheKey:
-    """Serves as a unique key for the discretization cache in
-    :meth:`GeometryCollection._get_cache`.
-    """
-
-
-class _GeometryCollectionConnectionCacheKey:
-    """Serves as a unique key for the connection cache in
-    :meth:`GeometryCollection._get_cache`.
-    """
-
-
-class GeometryCollection:
-    """A mapping from symbolic identifiers ("place IDs", typically strings)
-    to 'geometries', where a geometry can be a
-    :class:`~pytential.source.PotentialSource`, a
-    :class:`~pytential.target.TargetBase` or a
-    :class:`~meshmode.discretization.Discretization`.
-
-    This class is meant to hold a specific combination of sources and targets
-    serve to host caches of information derived from them, e.g. FMM trees
-    of subsets of them, as well as related common subexpressions such as
-    metric terms.
-
-    Refinement of :class:`pytential.qbx.QBXLayerPotentialSource` entries is
-    performed on demand, i.e. on calls to :meth:`get_discretization` with
-    a specific *discr_stage*. To perform refinement explicitly, call
-    :func:`pytential.qbx.refinement.refine_geometry_collection`,
-    which allows more customization of the refinement process through
-    parameters.
-
-    .. automethod:: __init__
-
-    .. attribute:: auto_source
-
-        Default :class:`~pytential.symbolic.primitives.DOFDescriptor` for the
-        source geometry.
-
-    .. attribute:: auto_target
-
-        Default :class:`~pytential.symbolic.primitives.DOFDescriptor` for the
-        target geometry.
-
-    .. automethod:: get_geometry
-    .. automethod:: get_discretization
-    .. automethod:: get_connection
-
-    .. automethod:: copy
-    .. automethod:: merge
-
-    """
-
-    def __init__(self, places, auto_where=None):
-        r"""
-        :arg places: a scalar, tuple of or mapping of symbolic names to
-            geometry objects. Supported objects are
-            :class:`~pytential.source.PotentialSource`,
-            :class:`~pytential.target.TargetBase` and
-            :class:`~meshmode.discretization.Discretization`. If this is
-            a mapping, the keys that are strings must be valid Python identifiers.
-            The tuple should contain only two entries, denoting the source and
-            target geometries for layer potential evaluation, identified by
-            *auto_where*.
-
-        :arg auto_where: a single or a tuple of two
-            :class:`~pytential.symbolic.primitives.DOFDescriptor`\ s, or values
-            that can be converted to one using
-            :func:`~pytential.symbolic.primitives.as_dofdesc`. The two
-            descriptors are used to define the default source and target
-            geometries for layer potential evaluations.
-            By default, they are set to
-            :class:`~pytential.symbolic.primitives.DEFAULT_SOURCE` and
-            :class:`~pytential.symbolic.primitives.DEFAULT_TARGET` for
-            sources and targets, respectively.
-        """
-
-        from pytential.target import TargetBase
-        from pytential.source import PotentialSource
-        from pytential.qbx import QBXLayerPotentialSource
-        from meshmode.discretization import Discretization
-
-        # {{{ construct dict
-
-        places_dict = {}
-
-        auto_source, auto_target = _prepare_auto_where(auto_where)
-        if isinstance(places, QBXLayerPotentialSource):
-            places_dict[auto_source.geometry] = places
-            auto_target = auto_source
-        elif isinstance(places, TargetBase):
-            places_dict[auto_target.geometry] = places
-            auto_source = auto_target
-        if isinstance(places, (Discretization, PotentialSource)):
-            places_dict[auto_source.geometry] = places
-            places_dict[auto_target.geometry] = places
-        elif isinstance(places, tuple):
-            source_discr, target_discr = places
-            places_dict[auto_source.geometry] = source_discr
-            places_dict[auto_target.geometry] = target_discr
-        else:
-            places_dict = places
-
-        import immutables
-        self.places = immutables.Map(places_dict)
-        self.auto_where = (auto_source, auto_target)
-
-        self._caches = {}
-
-        # }}}
-
-        # {{{ validate
-
-        # check auto_where
-        if auto_source.geometry not in self.places:
-            raise ValueError("'auto_where' source geometry is not in the "
-                f"collection: '{auto_source.geometry}'")
-
-        if auto_target.geometry not in self.places:
-            raise ValueError("'auto_where' target geometry is not in the "
-                f"collection: '{auto_target.geometry}'")
-
-        # check allowed identifiers
-        for name in self.places:
-            if not isinstance(name, str):
-                continue
-            if not _is_valid_identifier(name):
-                raise ValueError(f"'{name}' is not a valid identifier")
-
-        # check allowed types
-        for p in self.places.values():
-            if not isinstance(p, (PotentialSource, TargetBase, Discretization)):
-                raise TypeError(
-                    "Values in 'places' must be discretization, targets "
-                    f"or layer potential sources, got '{type(p).__name__}'")
-
-        # check ambient_dim
-        from pytools import is_single_valued
-        ambient_dims = [p.ambient_dim for p in self.places.values()]
-        if not is_single_valued(ambient_dims):
-            raise RuntimeError("All 'places' must have the same ambient dimension.")
-
-        self.ambient_dim = ambient_dims[0]
-
-        # }}}
-
-    @property
-    def auto_source(self):
-        return self.auto_where[0]
-
-    @property
-    def auto_target(self):
-        return self.auto_where[1]
-
-    # {{{ cache handling
-
-    def _get_cache(self, name):
-        return self._caches.setdefault(name, {})
-
-    def _get_discr_from_cache(self, geometry, discr_stage):
-        cache = self._get_cache(_GeometryCollectionDiscretizationCacheKey)
-        key = (geometry, discr_stage)
-
-        if key not in cache:
-            raise KeyError(
-                    f"cached discretization does not exist on '{geometry}' "
-                    f"for stage '{discr_stage}'")
-
-        return cache[key]
-
-    def _add_discr_to_cache(self, discr, geometry, discr_stage):
-        cache = self._get_cache(_GeometryCollectionDiscretizationCacheKey)
-        key = (geometry, discr_stage)
-
-        if key in cache:
-            raise RuntimeError("trying to overwrite the discretization cache of "
-                    f"'{geometry}' for stage '{discr_stage}'")
-
-        cache[key] = discr
-
-    def _get_conn_from_cache(self, geometry, from_stage, to_stage):
-        cache = self._get_cache(_GeometryCollectionConnectionCacheKey)
-        key = (geometry, from_stage, to_stage)
-
-        if key not in cache:
-            raise KeyError("cached connection does not exist on "
-                    f"'{geometry}' from stage '{from_stage}' to '{to_stage}'")
-
-        return cache[key]
-
-    def _add_conn_to_cache(self, conn, geometry, from_stage, to_stage):
-        cache = self._get_cache(_GeometryCollectionConnectionCacheKey)
-        key = (geometry, from_stage, to_stage)
-
-        if key in cache:
-            raise RuntimeError("trying to overwrite the connection cache of "
-                    f"'{geometry}' from stage '{from_stage}' to '{to_stage}'")
-
-        cache[key] = conn
-
-    def _get_qbx_discretization(self, geometry, discr_stage):
-        lpot_source = self.get_geometry(geometry)
-
-        try:
-            discr = self._get_discr_from_cache(geometry, discr_stage)
-        except KeyError:
-            from pytential.qbx.refinement import _refine_for_global_qbx
-
-            # NOTE: this adds the required discretizations to the cache
-            dofdesc = sym.DOFDescriptor(geometry, discr_stage)
-            _refine_for_global_qbx(self, dofdesc,
-                    lpot_source.refiner_code_container.get_wrangler(),
-                    _copy_collection=False)
-
-            discr = self._get_discr_from_cache(geometry, discr_stage)
-
-        return discr
-
-    # }}}
-
-    def get_connection(self, from_dd, to_dd):
-        """Construct a connection from *from_dd* to *to_dd* geometries.
-
-        :param from_dd: a :class:`~pytential.symbolic.primitives.DOFDescriptor`
-            or a value that can be converted to one using
-            :func:`~pytential.symbolic.primitives.as_dofdesc`.
-        :param to_dd: as *from_dd*.
-
-        :returns: an object compatible with the
-            :class:`~meshmode.discretization.connection.DiscretizationConnection`
-            interface.
-        """
-
-        from pytential.symbolic.dof_connection import connection_from_dds
-        return connection_from_dds(self, from_dd, to_dd)
-
-    def get_discretization(self, geometry, discr_stage=None):
-        """Get the geometry or discretization in the collection.
-
-        If a specific QBX stage discretization is requested, refinement is
-        performed on demand and cached for subsequent calls.
-
-        :param geometry: the identifier of the geometry in the collection.
-        :param discr_stage: if the geometry is a
-            :class:`~pytential.source.LayerPotentialSourceBase`, this denotes
-            the QBX stage of the returned discretization. Can be one of
-            :class:`~pytential.symbolic.primitives.QBX_SOURCE_STAGE1` (default),
-            :class:`~pytential.symbolic.primitives.QBX_SOURCE_STAGE2` or
-            :class:`~pytential.symbolic.primitives.QBX_SOURCE_QUAD_STAGE2`.
-
-        :returns: a geometry object in the collection or a
-            :class:`~meshmode.discretization.Discretization` corresponding to
-            *discr_stage*.
-        """
-        if discr_stage is None:
-            discr_stage = sym.QBX_SOURCE_STAGE1
-        discr = self.get_geometry(geometry)
-
-        from pytential.qbx import QBXLayerPotentialSource
-        from pytential.source import LayerPotentialSourceBase
-
-        if isinstance(discr, QBXLayerPotentialSource):
-            return self._get_qbx_discretization(geometry, discr_stage)
-        elif isinstance(discr, LayerPotentialSourceBase):
-            return discr.density_discr
-        else:
-            return discr
-
-    def get_geometry(self, geometry):
-        """
-        :param geometry: the identifier of the geometry in the collection.
-        """
-
-        try:
-            return self.places[geometry]
-        except KeyError:
-            raise KeyError(f"geometry not in the collection: '{geometry}'")
-
-    def copy(self, places=None, auto_where=None):
-        """Get a shallow copy of the geometry collection."""
-        return type(self)(
-                places=self.places if places is None else places,
-                auto_where=self.auto_where if auto_where is None else auto_where)
-
-    def merge(self, places):
-        """Merges two geometry collections and returns the new collection.
-
-        :param places: a :class:`dict` or :class:`GeometryCollection` to
-            merge with the current collection. If it is empty, a copy of the
-            current collection is returned.
-        """
-
-        new_places = self.places
-        if places:
-            if isinstance(places, GeometryCollection):
-                places = places.places
-            new_places = new_places.update(places)
-
-        return self.copy(places=new_places)
-
-    def __repr__(self):
-        return f"{type(self).__name__}({repr(self.places)})"
-
-    def __str__(self):
-        return f"{type(self).__name__}({repr(self.places)})"
-
-
-_KNOWN_GEOMETRY_COLLECTION_CACHE_KEYS = (
-        EvaluationMapperCSECacheKey,
-        _GeometryCollectionConnectionCacheKey,
-        _GeometryCollectionDiscretizationCacheKey,
-        )
-
-
-def add_geometry_to_collection(
-        places: GeometryCollection,
-        geometries: Dict[Hashable, Any]) -> GeometryCollection:
-    """Adds a :class:`dict` of geometries to an existing collection.
-
-    This function is similar to :meth:`GeometryCollection.merge`, but it makes
-    an attempt to maintain the caches in *places*. In particular, a shallow
-    copy of the following are pased to the new collection
-
-    * Any cached discretizations from
-      :func:`~pytential.qbx.refinement.refine_geometry_collection`.
-    * Any cached expressions marked with `cse_scope.DISCRETIZATION` from the
-      evaluation mapper.
-
-    This allows adding new targets to the collection without recomputing the
-    source geometry data.
-    """
-    from pytential.source import PointPotentialSource
-    from pytential.target import PointsTarget
-    for key, geometry in geometries.items():
-        if key in places.places:
-            raise ValueError(f"geometry '{key}' already in the collection")
-
-        if not isinstance(geometry, (PointsTarget, PointPotentialSource)):
-            raise TypeError(
-                    f"Cannot add a geometry of type '{type(geometry).__name__}' "
-                    "to the existing collection. Construct a new collection "
-                    "instead.")
-
-    # copy over all the caches
-    new_places = places.merge(geometries)
-    for key in places._caches:
-        if key not in _KNOWN_GEOMETRY_COLLECTION_CACHE_KEYS:
-            from warnings import warn
-            warn(f"GeometryCollection cache key '{key}' is not known and will "
-                    "be dropped in the new collection.")
-            continue
-
-        new_cache = new_places._get_cache(key)
-        new_cache.update(places._get_cache(key))
-
-    return new_places
 
 # }}}
 
@@ -1037,7 +671,7 @@ def _find_array_context_from_args_in_context(context, supplied_array_context=Non
 
 class BoundExpression:
     """An expression readied for evaluation by binding it to a
-    :class:`~pytential.GeometryCollection`.
+    :class:`~pytential.collection.GeometryCollection`.
 
     .. automethod :: cost_per_stage
     .. automethod :: cost_per_box
@@ -1216,7 +850,7 @@ class BoundExpression:
 
 def bind(places, expr, auto_where=None):
     """
-    :arg places: a :class:`pytential.GeometryCollection`.
+    :arg places: a :class:`pytential.collection.GeometryCollection`.
         Alternatively, any list or mapping that is a valid argument for its
         constructor can also be used.
     :arg auto_where: for simple source-to-self or source-to-target
@@ -1227,11 +861,12 @@ def bind(places, expr, auto_where=None):
         in the form of a :mod:`numpy` object array
     :returns: a :class:`pytential.symbolic.execution.BoundExpression`
     """
+    from pytential import GeometryCollection
     if not isinstance(places, GeometryCollection):
         places = GeometryCollection(places, auto_where=auto_where)
         auto_where = places.auto_where
-    expr = _prepare_expr(places, expr, auto_where=auto_where)
 
+    expr = _prepare_expr(places, expr, auto_where=auto_where)
     return BoundExpression(places, expr)
 
 # }}}
@@ -1273,7 +908,7 @@ def build_matrix(actx, places, exprs, input_exprs, domains=None,
         auto_where=None, context=None):
     """
     :arg actx: a :class:`~arraycontext.PyOpenCLArrayContext`.
-    :arg places: a :class:`pytential.GeometryCollection`.
+    :arg places: a :class:`pytential.collection.GeometryCollection`.
         Alternatively, any list or mapping that is a valid argument for its
         constructor can also be used.
     :arg exprs: an array of expressions corresponding to the output block
